Add devices module & fix docstrings
